+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Weapon properties should now longer be weirdly spaced out vertically.
 - Character's bigraphy and description should no longer be invisible.
 - Modifying a weapon's reload value in the character sheet should once again work.
+- Supreme Accuracy/Aptitude/Durability should once again work.
 
 ## [2.4.0.2.7.0] - 2033-11-30